requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.4
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
decorator==4.4.2
future==0.18.2
huepy==1.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
iniconfig==1.1.1
instabot==0.117.0
mock==4.0.2
moviepy==1.0.3
numpy==1.19.4
packaging==20.7
Pillow==8.0.1
pluggy==0.13.1
praw==7.1.0
prawcore==1.5.0
proglog==0.1.9
py==1.9.0
pycodestyle==2.6.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.1.2
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
requests-toolbelt==0.9.1
responses==0.12.1
schedule==0.6.0
six==1.15.0
toml==0.10.2
tqdm==4.54.1
update-checker==0.18.0
urllib3==1.26.2
websocket-client==0.57.0
